Large rework of how packet routing and PRN works.
Instead of generating a PRN per-hop we now use one PRN for the entire
packet lifetime. This means we don't have to worry about links dropping
packets in a multi-route scenario. It also allows for the removal of the
content PRN.

With this change the protocol diverges pretty significantly from NBP and
is no longer wire compatible.